Move sorted_robust and tabular_writer into pyomo.common
